+++ js/ingredients.js
@@ -142,27 +142,4 @@ $(document).on('click', '#btnImportCSV', function(event){
 
 });
 
-function get_stats(){
-	$.ajax({ 
-		url: pvOnlineAPI, 
-		dataType: 'json',
-		data: {
-			'request': 'getStats'
-		},
-		success: function (resp) {
-			$('#ingredientsTotal').html('<p>Ingredients online: <strong>' + resp.data.stats.Ingredients.count + '</strong></p>');
-			
-			$('#synonymsTotal').html('Synonyms online: <strong>' + resp.data.stats.Synonyms.count + '</strong>');
-			$('#composTotal').html('Compositions online: <strong>' + resp.data.stats.Compositions.count + '</strong>');
-			$('#ingredientsLimit').html('Import limited to: <strong>' + resp.data.stats.Ingredients.batchLimit + '</strong> materials');
 
-		},
-		error: function (request, status, error) {
-			$('#pv_online_imp_area').html('<div class="alert alert-danger mt-3"><i class="bi bi-exclamation-circle mx-2"></i></i>Unable to handle request, server returned an error: '+request.status+'</div>');
-			$('#btnImport').prop('disabled', true);
-		},
-		
-	});
-};
-
-get_stats();